07_streaming/pyproject.toml
[project]
name = "07-streaming"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]
